Weight restriction
Player can only take items if his carrying limit is more than the item's
weight and the item can be taken. Changed weight from float to double

+++ Item.java
@@ -10,7 +10,7 @@ public class Item
 {
     // instance variables - replace the example below with your own
     private String name;
-    private float weight;
+    private double weight;
     private boolean canTake;
 
     /**
@@ -25,13 +25,13 @@ public class Item
     public String getName(){
         return name;
     }
-    public float getWeight(){
+    public double getWeight(){
         return weight;
     }
     public boolean canTake(){
         return canTake;
     }
-    public void updateItem(String name, float weight, boolean canTake){
+    public void updateItem(String name, double weight, boolean canTake){
         this.name = name;
         this.weight = weight;
         this.canTake = canTake;
